Dropped debugger_middleware.
--HG--
branch : 0.7
rename : tipfy/debugger/app.py => tipfy/debugger/__init__.py
